Rename templates folder
