+++ cmd/templates/browser.js
