+++ cmd/templates/minimal.js
